+++ config.c
@@ -2433,6 +2433,9 @@ parse_section_tweak(struct context *ctx)
     else if (strcmp(key, "font-monospace-warn") == 0)
         return value_to_bool(ctx, &conf->tweak.font_monospace_warn);
 
+    else if (strcmp(key, "sixel") == 0)
+        return value_to_bool(ctx, &conf->tweak.sixel);
+
     else {
         LOG_CONTEXTUAL_ERR("not a valid option: %s", key);
         return false;
@@ -2974,6 +2977,7 @@ config_load(struct config *conf, const char *conf_path,
             .box_drawing_base_thickness = 0.04,
             .box_drawing_solid_shades = true,
             .font_monospace_warn = true,
+            .sixel = true,
         },
 
         .notifications = tll_init(),